content_bot_mvp/requirements.txt
aiogram>=3.13.0
aiosqlite
python-dotenv
openai
requests
pydantic-settings
